Update pt.json
Translation to portuguese (PT). This translation corrects some terms, verbal conjugations and phrase structure that are slightly different to the Brasilian norm. The ortography is that previous to the latest ortographical agreement thet is not yet madatorily in use unless by governmentally controlled organisations. I suggest oppening the possibility of using both ortographies as the application of that agreement is still a controversial issue.